refactor search tui
